Code review changes part 1
