+++ pkg/uhttp/wrapper.go
@@ -334,7 +334,7 @@ func (c *BaseHttpClient) Do(req *http.Request, options ...DoOption) (*http.Respo
 	}
 
 	if req.Method == http.MethodGet {
-		cacheKey, err = c.baseHttpCache.CreateCacheKey(req)
+		cacheKey, err = CreateCacheKey(req)
 		if err != nil {
 			return nil, err
 		}
